requirements.txt
streamlit
python-dotenv
google-genai
openai
pandas
matplotlib